removed structs to different file
