+++ BMP/bitmapMetadata.hpp
@@ -1,0 +1,29 @@
+#pragma once
+
+struct FileHeader {
+    short bfType;
+    int bfSize;
+    short bfReserved1;
+    short bfReserved2;
+    int bfOffBits;
+} File;
+
+struct PictureHeader {
+    int biSize;
+    int biWidth;
+    int biHeight;
+    short biPlanes;
+    short biBitCount;
+    int biCompression;
+    int biSizeImage;
+    int biXPelsPerMeter;
+    int biYPelsPerMeter;
+    int biClrUsed;
+    int biClrImportant;
+} Picture;
+
+struct ColorsRGB {
+    char R;
+    char G;
+    char B;
+} Rgb;
