+++ BMP/main.cpp
@@ -1,32 +1,5 @@
 #include <iostream>
-
-struct FileHeader {
-    short bfType;
-    int bfSize;
-    short bfReserved1;
-    short bfReserved2;
-    int bfOffBits;
-} File;
-
-struct PictureHeader {
-    int biSize;
-    int biWidth;
-    int biHeight;
-    short biPlanes;
-    short biBitCount;
-    int biCompression;
-    int biSizeImage;
-    int biXPelsPerMeter;
-    int biYPelsPerMeter;
-    int biClrUsed;
-    int biClrImportant;
-} Picture;
-
-struct ColorsRGB {
-    char R;
-    char G;
-    char B;
-} Rgb;
+#include "bitmapMetadata.hpp"
 
 int main(int arc, char* argv[])
 {
